# remove: db index for validity-prover's leaves table