--- a/validity-prover/migrations/20250307095617_del_index_for_leaves.down.sql
+++ b/validity-prover/migrations/20250307095617_del_index_for_leaves.down.sql
@@ -1,0 +1,2 @@
+CREATE INDEX idx_leaves_lookup ON public.leaves USING btree ("position", tag, timestamp_value DESC);
+CREATE INDEX idx_leaves_timestamp ON public.leaves USING btree (timestamp_value DESC, tag);