--- a/validity-prover/migrations/20250307095617_del_index_for_leaves.up.sql
+++ b/validity-prover/migrations/20250307095617_del_index_for_leaves.up.sql
@@ -1,0 +1,2 @@
+DROP INDEX IF EXISTS idx_leaves_lookup;
+DROP INDEX IF EXISTS idx_leaves_timestamp;