## Change Summary:
This PR adds migration scripts to remove two indexes ( `idx_leaves_lookup` and `idx_leaves_timestamp` ) from the leaves table. These indexes are not being effectively utilized in our current query patterns and are being removed to improve database performance.