@xbicsan 2025-03-09
# benchmark results

benchmark script:
```
diff --git a/validity-prover/src/trees/merkle_tree/mod.rs b/validity-prover/src/trees/merkle_tree/mod.rs
index e380d2b..1ff928e 100644
--- a/validity-prover/src/trees/merkle_tree/mod.rs
+++ b/validity-prover/src/trees/merkle_tree/mod.rs
@@ -94,10 +94,10 @@ mod tests {
     type V = u32;

     #[tokio::test]
-    #[ignore]
+    //#[ignore]
     async fn test_speed_incremental_merkle_tree() -> anyhow::Result<()> {
         let height = 32;
-        let n = 1 << 8;
+        let n = 1000;
         let mut rng = rand::thread_rng();

         let database_url = setup_test();
@@ -106,18 +106,44 @@ mod tests {
         let tree = SqlIncrementalMerkleTree::<V>::new(pool, rng.gen(), height);
         tree.reset(0).await?;

-        let timestamp = 0;
         let time = std::time::Instant::now();
         for i in 0..n {
+            let timestamp = i;
             tree.push(timestamp, i as u32).await?;
         }
         println!(
-            "SqlIncrementMerkleTree: {} leaves, {} height, {} seconds",
+            "SqlIncrementMerkleTree.push: {} leaves, {} height, {} seconds",
             n,
             height,
             time.elapsed().as_secs_f64()
         );

+        let n = n * 5;
+
+        // SELECT - pattern 1
+        let time = std::time::Instant::now();
+        for _i in 0..n {
+            tree.get_last_timestamp().await?;
+        }
+        println!(
+            "SqlIncrementMerkleTree.get_last_timestamp: {} times, {} seconds",
+            n,
+            time.elapsed().as_secs_f64()
+        );
+
+        // SELECT - pattern 2
+        let time = std::time::Instant::now();
+        for _i in 0..n {
+            tree.get_leaf(n * 2, 1).await?;
+            tree.get_leaf(n * 2, n / 4).await?;
+            tree.get_leaf(n * 2, n / 2).await?;
+        }
+        println!(
+            "SqlIncrementMerkleTree.get_leaf: {} times, {} seconds",
+            n,
+            time.elapsed().as_secs_f64()
+        );
+
         Ok(())
     }
```

## Before
before execution benchmark:
```
 schema_name |             table_name              |                index_name                | scans | index_size | total_size | index_ratio | tuples_read | tuples_fetched
-------------+-------------------------------------+------------------------------------------+-------+------------+------------+-------------+-------------+----------------
 public      | _sqlx_migrations                    | _sqlx_migrations_pkey                    |     2 | 16 kB      | 208 kB     |        7.69 |           6 |              6
 public      | observer_block_sync_eth_block_num   | observer_block_sync_eth_block_num_pkey   |     0 | 8192 bytes | 208 kB     |        3.85 |           0 |              0
 public      | observer_deposit_sync_eth_block_num | observer_deposit_sync_eth_block_num_pkey |     0 | 8192 bytes | 208 kB     |        3.85 |           0 |              0
 public      | full_blocks                         | full_blocks_pkey                         |     0 | 8192 bytes | 208 kB     |        3.85 |           0 |              0
 public      | deposit_leaf_events                 | deposit_leaf_events_pkey                 |     0 | 8192 bytes | 208 kB     |        3.85 |           0 |              0
 public      | validity_state                      | validity_state_pkey                      |     0 | 8192 bytes | 208 kB     |        3.85 |           0 |              0
 public      | tx_tree_roots                       | tx_tree_roots_pkey                       |     0 | 8192 bytes | 208 kB     |        3.85 |           0 |              0
 public      | validity_proofs                     | validity_proofs_pkey                     |     0 | 8192 bytes | 208 kB     |        3.85 |           0 |              0
 public      | prover_tasks                        | prover_tasks_pkey                        |     0 | 8192 bytes | 208 kB     |        3.85 |           0 |              0
 public      | hash_nodes                          | hash_nodes_pkey                          |     0 | 8192 bytes | 208 kB     |        3.85 |           0 |              0
 public      | leaves                              | leaves_pkey                              |     0 | 8192 bytes | 208 kB     |        3.85 |           0 |              0
 public      | leaves_len                          | leaves_len_pkey                          |     0 | 8192 bytes | 208 kB     |        3.85 |           0 |              0
 public      | deposit_leaf_events                 | idx_deposit_leaf_events_deposit_hash     |     0 | 8192 bytes | 208 kB     |        3.85 |           0 |              0
 public      | deposit_leaf_events                 | idx_deposit_leaf_events_block_tx         |     0 | 8192 bytes | 208 kB     |        3.85 |           0 |              0
 public      | full_blocks                         | idx_full_blocks_block_tx                 |     0 | 8192 bytes | 208 kB     |        3.85 |           0 |              0
 public      | hash_nodes                          | idx_hash_nodes_timestamp                 |     0 | 8192 bytes | 208 kB     |        3.85 |           0 |              0
 public      | hash_nodes                          | idx_hash_nodes_lookup                    |     0 | 8192 bytes | 208 kB     |        3.85 |           0 |              0
 public      | leaves_len                          | idx_leaves_len_lookup                    |     0 | 8192 bytes | 208 kB     |        3.85 |           0 |              0
 public      | indexed_leaves                      | indexed_leaves_pkey                      |     0 | 8192 bytes | 208 kB     |        3.85 |           0 |              0
 public      | indexed_leaves                      | idx_indexed_leaves_lookup                |     0 | 8192 bytes | 208 kB     |        3.85 |           0 |              0
 public      | indexed_leaves                      | idx_indexed_leaves_timestamp             |     0 | 8192 bytes | 208 kB     |        3.85 |           0 |              0
 public      | prover_tasks                        | idx_prover_tasks_assigned_status         |     0 | 8192 bytes | 208 kB     |        3.85 |           0 |              0
 public      | tx_tree_roots                       | idx_tx_tree_roots_block_number           |     0 | 8192 bytes | 208 kB     |        3.85 |           0 |              0
 public      | leaves                              | idx_leaves_lookup                        |     0 | 8192 bytes | 208 kB     |        3.85 |           0 |              0
 public      | leaves                              | idx_leaves_timestamp                     |     0 | 8192 bytes | 208 kB     |        3.85 |           0 |              0
(25 rows)
```

execute benchmark:
```
$ cargo test -r trees::merkle_tree::tests::test_speed_incremental_merkle_tree -- --nocapture
running 1 test
SqlIncrementMerkleTree.push: 1000 leaves, 32 height, 21.558983542 seconds
SqlIncrementMerkleTree.get_last_timestamp: 5000 times, 7.139009125 seconds
SqlIncrementMerkleTree.get_leaf: 5000 times, 21.423146125 seconds
test trees::merkle_tree::tests::test_speed_incremental_merkle_tree ... ok

test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 7 filtered out; finished in 50.15s
```

after execution benchmark:
```
 schema_name |             table_name              |                index_name                | scans | index_size | total_size | index_ratio | tuples_read | tuples_fetched
-------------+-------------------------------------+------------------------------------------+-------+------------+------------+-------------+-------------+----------------
 public      | hash_nodes                          | hash_nodes_pkey                          | 33000 | 1616 kB    | 5104 kB    |       31.66 |           0 |              0
 public      | hash_nodes                          | idx_hash_nodes_lookup                    | 32000 | 2576 kB    | 5104 kB    |       50.47 |        4932 |           4932
 public      | leaves                              | idx_leaves_lookup                        | 15000 | 56 kB      | 5104 kB    |        1.10 |        5000 |           5000
 public      | leaves                              | leaves_pkey                              |  6000 | 56 kB      | 5104 kB    |        1.10 |        5000 |           5000
 public      | leaves_len                          | idx_leaves_len_lookup                    |  2001 | 72 kB      | 5104 kB    |        1.41 |        1998 |           1998
 public      | leaves_len                          | leaves_len_pkey                          |  1000 | 48 kB      | 5104 kB    |        0.94 |           0 |              0
 public      | _sqlx_migrations                    | _sqlx_migrations_pkey                    |     2 | 16 kB      | 5104 kB    |        0.31 |           6 |              6
 public      | hash_nodes                          | idx_hash_nodes_timestamp                 |     1 | 464 kB     | 5104 kB    |        9.09 |           0 |              0
 public      | leaves                              | idx_leaves_timestamp                     |     1 | 72 kB      | 5104 kB    |        1.41 |           0 |              0
 public      | prover_tasks                        | prover_tasks_pkey                        |     0 | 8192 bytes | 5104 kB    |        0.16 |           0 |              0
 public      | validity_proofs                     | validity_proofs_pkey                     |     0 | 8192 bytes | 5104 kB    |        0.16 |           0 |              0
 public      | tx_tree_roots                       | tx_tree_roots_pkey                       |     0 | 8192 bytes | 5104 kB    |        0.16 |           0 |              0
 public      | deposit_leaf_events                 | idx_deposit_leaf_events_deposit_hash     |     0 | 8192 bytes | 5104 kB    |        0.16 |           0 |              0
 public      | deposit_leaf_events                 | idx_deposit_leaf_events_block_tx         |     0 | 8192 bytes | 5104 kB    |        0.16 |           0 |              0
 public      | full_blocks                         | idx_full_blocks_block_tx                 |     0 | 8192 bytes | 5104 kB    |        0.16 |           0 |              0
 public      | validity_state                      | validity_state_pkey                      |     0 | 8192 bytes | 5104 kB    |        0.16 |           0 |              0
 public      | deposit_leaf_events                 | deposit_leaf_events_pkey                 |     0 | 8192 bytes | 5104 kB    |        0.16 |           0 |              0
 public      | full_blocks                         | full_blocks_pkey                         |     0 | 8192 bytes | 5104 kB    |        0.16 |           0 |              0
 public      | indexed_leaves                      | indexed_leaves_pkey                      |     0 | 8192 bytes | 5104 kB    |        0.16 |           0 |              0
 public      | indexed_leaves                      | idx_indexed_leaves_lookup                |     0 | 8192 bytes | 5104 kB    |        0.16 |           0 |              0
 public      | indexed_leaves                      | idx_indexed_leaves_timestamp             |     0 | 8192 bytes | 5104 kB    |        0.16 |           0 |              0
 public      | prover_tasks                        | idx_prover_tasks_assigned_status         |     0 | 8192 bytes | 5104 kB    |        0.16 |           0 |              0
 public      | tx_tree_roots                       | idx_tx_tree_roots_block_number           |     0 | 8192 bytes | 5104 kB    |        0.16 |           0 |              0
 public      | observer_deposit_sync_eth_block_num | observer_deposit_sync_eth_block_num_pkey |     0 | 8192 bytes | 5104 kB    |        0.16 |           0 |              0
 public      | observer_block_sync_eth_block_num   | observer_block_sync_eth_block_num_pkey   |     0 | 8192 bytes | 5104 kB    |        0.16 |           0 |              0
(25 rows)
```

## After
before execution benchmark:
```
 schema_name |             table_name              |                index_name                | scans | index_size | total_size | index_ratio | tuples_read | tuples_fetched
-------------+-------------------------------------+------------------------------------------+-------+------------+------------+-------------+-------------+----------------
 public      | _sqlx_migrations                    | _sqlx_migrations_pkey                    |     0 | 16 kB      | 192 kB     |        8.33 |           0 |              0
 public      | observer_block_sync_eth_block_num   | observer_block_sync_eth_block_num_pkey   |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | observer_deposit_sync_eth_block_num | observer_deposit_sync_eth_block_num_pkey |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | full_blocks                         | full_blocks_pkey                         |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | deposit_leaf_events                 | deposit_leaf_events_pkey                 |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | validity_state                      | validity_state_pkey                      |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | tx_tree_roots                       | tx_tree_roots_pkey                       |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | validity_proofs                     | validity_proofs_pkey                     |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | prover_tasks                        | prover_tasks_pkey                        |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | hash_nodes                          | hash_nodes_pkey                          |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | leaves                              | leaves_pkey                              |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | leaves_len                          | leaves_len_pkey                          |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | deposit_leaf_events                 | idx_deposit_leaf_events_deposit_hash     |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | deposit_leaf_events                 | idx_deposit_leaf_events_block_tx         |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | full_blocks                         | idx_full_blocks_block_tx                 |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | hash_nodes                          | idx_hash_nodes_timestamp                 |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | hash_nodes                          | idx_hash_nodes_lookup                    |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | leaves_len                          | idx_leaves_len_lookup                    |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | indexed_leaves                      | indexed_leaves_pkey                      |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | indexed_leaves                      | idx_indexed_leaves_lookup                |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | indexed_leaves                      | idx_indexed_leaves_timestamp             |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | prover_tasks                        | idx_prover_tasks_assigned_status         |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
 public      | tx_tree_roots                       | idx_tx_tree_roots_block_number           |     0 | 8192 bytes | 192 kB     |        4.17 |           0 |              0
(23 rows)
```

execute benchmark:
```
$ cargo test -r trees::merkle_tree::tests::test_speed_incremental_merkle_tree -- --nocapture
running 1 test
SqlIncrementMerkleTree.push: 1000 leaves, 32 height, 20.994767458 seconds
SqlIncrementMerkleTree.get_last_timestamp: 5000 times, 6.993182333 seconds
SqlIncrementMerkleTree.get_leaf: 5000 times, 21.959066667 seconds
test trees::merkle_tree::tests::test_speed_incremental_merkle_tree ... ok

test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 7 filtered out; finished in 49.97s
```

after execution benchmark:
```
 schema_name |             table_name              |                index_name                | scans | index_size | total_size | index_ratio | tuples_read | tuples_fetched
-------------+-------------------------------------+------------------------------------------+-------+------------+------------+-------------+-------------+----------------
 public      | hash_nodes                          | hash_nodes_pkey                          | 33000 | 1616 kB    | 4976 kB    |       32.48 |           0 |              0
 public      | hash_nodes                          | idx_hash_nodes_lookup                    | 32000 | 2576 kB    | 4976 kB    |       51.77 |        4932 |           4932
 public      | leaves                              | leaves_pkey                              | 21001 | 56 kB      | 4976 kB    |        1.13 |       10006 |          10006
 public      | leaves_len                          | idx_leaves_len_lookup                    |  1701 | 72 kB      | 4976 kB    |        1.45 |        1698 |           1698
 public      | leaves_len                          | leaves_len_pkey                          |  1300 | 48 kB      | 4976 kB    |        0.96 |         300 |            300
 public      | hash_nodes                          | idx_hash_nodes_timestamp                 |     1 | 464 kB     | 4976 kB    |        9.32 |           0 |              0
 public      | _sqlx_migrations                    | _sqlx_migrations_pkey                    |     0 | 16 kB      | 4976 kB    |        0.32 |           0 |              0
 public      | validity_proofs                     | validity_proofs_pkey                     |     0 | 8192 bytes | 4976 kB    |        0.16 |           0 |              0
 public      | prover_tasks                        | prover_tasks_pkey                        |     0 | 8192 bytes | 4976 kB    |        0.16 |           0 |              0
 public      | tx_tree_roots                       | tx_tree_roots_pkey                       |     0 | 8192 bytes | 4976 kB    |        0.16 |           0 |              0
 public      | validity_state                      | validity_state_pkey                      |     0 | 8192 bytes | 4976 kB    |        0.16 |           0 |              0
 public      | deposit_leaf_events                 | deposit_leaf_events_pkey                 |     0 | 8192 bytes | 4976 kB    |        0.16 |           0 |              0
 public      | deposit_leaf_events                 | idx_deposit_leaf_events_deposit_hash     |     0 | 8192 bytes | 4976 kB    |        0.16 |           0 |              0
 public      | deposit_leaf_events                 | idx_deposit_leaf_events_block_tx         |     0 | 8192 bytes | 4976 kB    |        0.16 |           0 |              0
…